requirements.txt
azure-identity
azure-mgmt-costmanagement
azure-core
requests
pandas
pyyaml
numpy
prophet
statsmodels
openpyxl
